♻️ useAuthFormInitializeを作成
